# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Book Tracker API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type BookStatus = typeof BookStatus[keyof typeof BookStatus];


export const BookStatus = {
  reading: 'reading',
  read: 'read',
  want_to_read: 'want_to_read',
} as const;

export interface Book {
  id: string;
  title: string;
  author: string;
  coverColor: string;
  status: BookStatus;
  /** @nullable */
  rating?: number | null;
  /** @nullable */
  pages?: number | null;
  /** @nullable */
  currentPage?: number | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  genre?: string | null;
  dateAdded: string;
  /** @nullable */
  dateStarted?: string | null;
  /** @nullable */
  dateFinished?: string | null;
}

export type BookInputStatus = typeof BookInputStatus[keyof typeof BookInputStatus];


export const BookInputStatus = {
  reading: 'reading',
  read: 'read',
  want_to_read: 'want_to_read',
} as const;

export interface BookInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  author: string;
  status: BookInputStatus;
  /** @nullable */
  rating?: number | null;
  /** @nullable */
  pages?: number | null;
  /** @nullable */
  currentPage?: number | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  genre?: string | null;
}

export type BookPatchStatus = typeof BookPatchStatus[keyof typeof BookPatchStatus];


export const BookPatchStatus = {
  reading: 'reading',
  read: 'read',
  want_to_read: 'want_to_read',
} as const;

export interface BookPatch {
  title?: string;
  author?: string;
  status?: BookPatchStatus;
  /** @nullable */
  rating?: number | null;
  /** @nullable */
  pages?: number | null;
  /** @nullable */
  currentPage?: number | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  genre?: string | null;
  /** @nullable */
  dateStarted?: string | null;
  /** @nullable */
  dateFinished?: string | null;
}

export interface GenreCount {
  genre: string;
  count: number;
}

export interface BookStats {
  total: number;
  reading: number;
  read: number;
  wantToRead: number;
  readThisYear: number;
  totalPages: number;
  /** @nullable */
  avgRating: number | null;
  topGenres: GenreCount[];
}

export interface EnrichAllResult {
  enriching: number;
}

export interface BookLookupResult {
  title: string;
  author: string;
  /** @nullable */
  pages?: number | null;
  /** @nullable */
  genre?: string | null;
  /** @nullable */
  coverUrl?: string | null;
  /** @nullable */
  publishYear?: number | null;
}

export type ListBooksParams = {
status?: ListBooksStatus;
};

export type ListBooksStatus = typeof ListBooksStatus[keyof typeof ListBooksStatus];


export const ListBooksStatus = {
  reading: 'reading',
  read: 'read',
  want_to_read: 'want_to_read',
} as const;

export type LookupBookByIsbnParams = {
isbn: string;
};